chainlit-sample-db/requirements.txt
chainlit>=1.0.0
openai>=1.6.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
greenlet>=2.0.0
